Refactored roles to support additional features
To allow a way to call a description message for each role, I've set the role meta data to static and added an underlying protected variable for overriding on a player by player basis. This will allow roles to have their teams or appearances changed on the fly. Role base class has function to print the description data to the game channel.

Other misc. changes:
- Moved the thumbnail data into the class it is referenced by to reduce the amount of code with confusing context.
- Added the Cult Green color for the Cult team when it gets implemented.
- Witch night action rules includes additional information no players they save.
- ESLINT rules for max-length count do not apply to code with strings in them.